+++ src/sns/report/model/mod.rs
@@ -3,7 +3,19 @@ mod reports;
 mod requests;
 mod sorts;
 
-pub use errors::*;
-pub use reports::*;
-pub use requests::*;
-pub use sorts::*;
+pub use errors::SnsHostError;
+#[cfg(test)]
+pub use reports::{SnsCustomProposalCriticality, SnsVotingRewardsParameters};
+pub use reports::{
+    SnsGovernanceParameters, SnsInfoReport, SnsListReport, SnsListRow, SnsNeuronPermissionList,
+    SnsNeuronRow, SnsNeuronsCacheListReport, SnsNeuronsCacheStatusReport, SnsNeuronsCacheSummary,
+    SnsNeuronsRefreshAttemptStatus, SnsNeuronsRefreshReport, SnsNeuronsReport, SnsParamsReport,
+    SnsProposalBallotRow, SnsProposalFailureReason, SnsProposalReport, SnsProposalRow,
+    SnsProposalTally, SnsProposalsReport, SnsTokenMetadataRow, SnsTokenReport, SnsTokenStandardRow,
+};
+pub use requests::{
+    SnsInfoRequest, SnsListRequest, SnsLookupRequest, SnsNeuronsCacheListRequest,
+    SnsNeuronsCacheStatusRequest, SnsNeuronsRefreshRequest, SnsNeuronsRequest, SnsParamsRequest,
+    SnsProposalRequest, SnsProposalsRequest, SnsTokenRequest,
+};
+pub use sorts::{SnsListSort, SnsNeuronsSort, SnsProposalStatusFilter};