+++ src/sns/report/source/mod.rs
@@ -1,5 +1,11 @@
 mod model;
 mod traits;
 
-pub(in crate::sns::report) use model::*;
-pub(in crate::sns::report) use traits::*;
+pub(in crate::sns::report) use model::{
+    MainnetSns, MainnetSnsCanisters, MainnetSnsList, MainnetSnsNeuronPage, MainnetSnsNeurons,
+    MainnetSnsProposal, MainnetSnsProposals, MainnetSnsToken, SnsFetchRequest, SnsNeuronId,
+};
+pub(in crate::sns::report) use traits::{
+    SnsListSource, SnsNeuronsSource, SnsParamsSource, SnsProposalSource, SnsProposalsSource,
+    SnsTokenSource,
+};